// src/app/com/annaniks/yev-yev/core/models/route-price.ts
export interface IRoutePrice {
    id?: number;
    sub_route: number;
    sunday_price: number;
    sunday_time: string;
    monday_price: number;
    monday_time: string;
    tuesday_price: number;
    tuesday_time: string;
    wednesday_price: number;
    wednesday_time: string;
    thursday_price: number;
    thursday_time: string;
    friday_price: number;
    friday_time: string;
    saturday_price: number;
    saturday_time: string;
    is_active: boolean;
}